// src/advertising/dto/credentials.sp.dto.ts
export interface CredentialsSpDto {
  access_token: string;
  refresh_token: string;
  seller_name: string;
  client_id: string;
  client_secret: string;
  access_key: string;
  secret_key: string;
  role_arn: string;
  selling_partner_id: string;
  region: string;
  marketPlaceName: string;
  marketPlaceId: string;
}