// src/advertising/enum/report.enum.ts
export enum ReportEnum {
  _GET_FLAT_FILE_OPEN_LISTINGS_DATA_ = '_GET_FLAT_FILE_OPEN_LISTINGS_DATA_',
  _GET_MERCHANT_LISTINGS_DATA_ = '_GET_MERCHANT_LISTINGS_DATA_',
  GET_MERCHANT_LISTINGS_ALL_DATA = 'GET_MERCHANT_LISTINGS_ALL_DATA',
  _GET_MERCHANT_CANCELLED_LISTINGS_DATA_ = '_GET_MERCHANT_CANCELLED_LISTINGS_DATA_',
  _GET_CONVERGED_FLAT_FILE_SOLD_LISTINGS_DATA_ = '_GET_CONVERGED_FLAT_FILE_SOLD_LISTINGS_DATA_',
  _GET_MERCHANT_LISTINGS_DEFECT_DATA_ = '_GET_MERCHANT_LISTINGS_DEFECT_DATA_',
  _GET_FLAT_FILE_ACTIONABLE_ORDER_DATA_ = '_GET_FLAT_FILE_ACTIONABLE_ORDER_DATA_',
  _GET_FLAT_FILE_ORDERS_DATA_ = '_GET_FLAT_FILE_ORDERS_DATA_',
  _GET_CONVERGED_FLAT_FILE_ORDER_REPORT_DATA_ = '_GET_CONVERGED_FLAT_FILE_ORDER_REPORT_DATA_',
  _GET_FLAT_FILE_ALL_ORDERS_DATA_BY_LAST_UPDATE_ = '_GET_FLAT_FILE_ALL_ORDERS_DATA_BY_LAST_UPDATE_',
  GET_FLAT_FILE_ALL_ORDERS_DATA_BY_ORDER_DATE_GENERAL = 'GET_FLAT_FILE_ALL_ORDERS_DATA_BY_ORDER_DATE_GENERAL',
  _GET_FBA_FULFILLMENT_REMOVAL_ORDER_DETAIL_DATA_ = '_GET_FBA_FULFILLMENT_REMOVAL_ORDER_DETAIL_DATA_',
  _GET_AMAZON_FULFILLED_SHIPMENTS_DATA_ = '_GET_AMAZON_FULFILLED_SHIPMENTS_DATA_',
  _GET_FBA_FULFILLMENT_CUSTOMER_SHIPMENT_SALES_DATA_ = '_GET_FBA_FULFILLMENT_CUSTOMER_SHIPMENT_SALES_DATA_',
  GET_FBA_FULFILLMENT_CUSTOMER_SHIPMENT_PROMOTION_DATA = 'GET_FBA_FULFILLMENT_CUSTOMER_SHIPMENT_PROMOTION_DATA',
  _GET_AFN_INVENTORY_DATA_ = '_GET_AFN_INVENTORY_DATA_',
  _GET_AFN_INVENTORY_DATA_BY_COUNTRY_ = '_GET_AFN_INVENTORY_DATA_BY_COUNTRY_',
  _GET_FBA_FULFILLMENT_INVENTORY_RECEIPTS_DATA_ = '_GET_FBA_FULFILLMENT_INVENTORY_RECEIPTS_DATA_',
  GET_RESERVED_INVENTORY_DATA = 'GET_RESERVED_INVENTORY_DATA',
  _GET_FBA_FULFILLMENT_MONTHLY_INVENTORY_DATA_ = '_GET_FBA_FULFILLMENT_MONTHLY_INVENTORY_DATA_',
  _GET_FBA_FULFILLMENT_INVENTORY_HEALTH_DATA_ = '_GET_FBA_FULFILLMENT_INVENTORY_HEALTH_DATA_',
  _GET_RESTOCK_INVENTORY_RECOMMENDATIONS_REPORT_ = '_GET_RESTOCK_INVENTORY_RECOMMENDATIONS_REPORT_',
  _GET_FBA_FULFILLMENT_INBOUND_NONCOMPLIANCE_DATA_ = '_GET_FBA_FULFILLMENT_INBOUND_NONCOMPLIANCE_DATA_',
  _GET_FBA_MYI_UNSUPPRESSED_INVENTORY_DATA_ = '_GET_FBA_MYI_UNSUPPRESSED_INVENTORY_DATA_',
  GET_FBA_MYI_ALL_INVENTORY_DATA = 'GET_FBA_MYI_ALL_INVENTORY_DATA',
  _GET_FBA_UNO_INVENTORY_DATA_ = '_GET_FBA_UNO_INVENTORY_DATA_',
  _GET_FBA_FULFILLMENT_INVENTORY_ADJUSTMENTS_DATA_ = '_GET_FBA_FULFILLMENT_INVENTORY_ADJUSTMENTS_DATA_',
  _GET_STRANDED_INVENTORY_UI_DATA_ = '_GET_STRANDED_INVENTORY_UI_DATA_',
  _GET_FBA_FULFILLMENT_INVENTORY_SUMMARY_DATA_ = '_GET_FBA_FULFILLMENT_INVENTORY_SUMMARY_DATA_',
  _GET_FBA_ESTIMATED_FBA_FEES_TXT_DATA_ = '_GET_FBA_ESTIMATED_FBA_FEES_TXT_DATA_',
  _GET_DATE_RANGE_FINANCIAL_TRANSACTION_DATA_ = '_GET_DATE_RANGE_FINANCIAL_TRANSACTION_DATA_',
  GET_V2_SETTLEMENT_REPORT_DATA_FLAT_FILE = 'GET_V2_SETTLEMENT_REPORT_DATA_FLAT_FILE',
  GET_FBA_REIMBURSEMENTS_DATA = 'GET_FBA_REIMBURSEMENTS_DATA',
  GET_FBA_FULFILLMENT_CUSTOMER_RETURNS_DATA = 'GET_FBA_FULFILLMENT_CUSTOMER_RETURNS_DATA',
  _GET_FBA_FULFILLMENT_CUSTOMER_SHIPMENT_REPLACEMENT_DATA_ = '_GET_FBA_FULFILLMENT_CUSTOMER_SHIPMENT_REPLACEMENT_DATA_',
  _GET_SELLER_FEEDBACK_DATA_ = '_GET_SELLER_FEEDBACK_DATA_',
  _GET_V1_SELLER_PERFORMANCE_REPORT_ = '_GET_V1_SELLER_PERFORMANCE_REPORT_',
  GET_SALES_AND_TRAFFIC_REPORT = 'GET_SALES_AND_TRAFFIC_REPORT',
  GET_REFERRAL_FEE_PREVIEW_REPORT = 'GET_REFERRAL_FEE_PREVIEW_REPORT',
}